SMHE-2412: Add object config for GetPeriodicMeterReadsCommandExecutor
