+++ osgp/shared/osgp-dto/src/main/java/org/opensmartgridplatform/dto/valueobjects/smartmetering/MeterReadsResponseDto.java
@@ -6,20 +6,14 @@ package org.opensmartgridplatform.dto.valueobjects.smartmetering;
 
 import java.util.Date;
 
-public class MeterReadsResponseDto extends ActionResponseDto {
+public class MeterReadsResponseDto extends MeterReadsResponseWithLogTimeDto {
   private static final long serialVersionUID = -297320204916085999L;
-
-  private final Date logTime;
 
   private final ActiveEnergyValuesDto activeEnergyValues;
 
   public MeterReadsResponseDto(final Date logTime, final ActiveEnergyValuesDto activeEnergyValues) {
-    this.logTime = new Date(logTime.getTime());
+    super(logTime);
     this.activeEnergyValues = activeEnergyValues;
-  }
-
-  public Date getLogTime() {
-    return new Date(this.logTime.getTime());
   }
 
   public DlmsMeterValueDto getActiveEnergyImportTariffOne() {
@@ -53,7 +47,7 @@ public class MeterReadsResponseDto extends ActionResponseDto {
   @Override
   public String toString() {
     return "MeterReads[logTime="
-        + this.logTime
+        + this.getLogTime()
         + ", "
         + this.activeEnergyValues
         + ", activeEnergyExport="
